backend/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
pdfplumber==0.11.4
google-generativeai==0.8.3
python-dotenv==1.0.1
pytesseract==0.3.13
pdf2image==1.17.0
pillow==10.4.0
pydantic-settings==2.5.2
pytest==8.3.3